Fix REST spotlight endpoint, add same logic of front to the backend, to search users and channels with # and @
