+++ packages/rocketchat-api/package.js
@@ -44,5 +44,4 @@ Package.onUse(function(api) {
 	api.addFiles('server/v1/settings.js', 'server');
 	api.addFiles('server/v1/stats.js', 'server');
 	api.addFiles('server/v1/users.js', 'server');
-	api.addFiles('server/v1/spotlight.js', 'server');
 });
